+++ src/Core/components/CfEddbkWizard.js
@@ -120,7 +120,14 @@ export default function (store, bookingDataMap, TranslateCapable, MapBookingFiel
          *
          * @property {string} errorMessage Error message if booking is not created.
          */
-        errorMessage: null
+        errorMessage: null,
+
+        /**
+         * @since [*next-version*]
+         *
+         * @property {object} filterValues Selected booking filters.
+         */
+        filterValues: {}
       }
     },
 
@@ -238,6 +245,17 @@ export default function (store, bookingDataMap, TranslateCapable, MapBookingFiel
 
     methods: {
       /**
+       * Update filter values.
+       *
+       * @since [*next-version*]
+       *
+       * @param values
+       */
+      updateFilterValues (values) {
+        this.filterValues = values
+      },
+
+      /**
        * Create booking on server and redirect user to cart, if url is
        * passed to component's config.
        *
@@ -251,7 +269,7 @@ export default function (store, bookingDataMap, TranslateCapable, MapBookingFiel
           start: this.session.start,
           end: this.session.end,
           service: this.service.id,
-          resource: this.session.resource,
+          resources: this.session.resources,
           transition: this.configuration.initialBookingTransition,
         }, this.bookingState)).then((response) => {
           this.isCreatingBooking = false